requirements.txt
python-dotenv
mysqlclient
fastapi
uvicorn
python-multipart